l22 Django/course22/requirements.txt
Django>=2.0
django-debug-toolbar>=1.9.1
pytz
sqlparse
wheel
django-queryinspect>=1.0.0

djangorestframework==3.7.3
raven
